refactors translation utils
