+++ packages/cms/src/api/cmsRoute.js
@@ -6,8 +6,8 @@ const fs = require("fs");
 const Op = sequelize.Op;
 
 const populateSearch = require("../utils/populateSearch");
-const translateContent = require("../utils/translateContent");
-const formatters4eval = require("../utils/formatters4eval");
+const {profileReqFull, storyReqFull, sectionReqFull, storysectionReqFull, cmsTables, contentTables, parentOrderingTables} = require("../utils/sequelize/models");
+const {upsertTranslation, translateSection} = require("../utils/translation/translationUtils");
 
 const envLoc = process.env.CANON_LANGUAGE_DEFAULT || "en";
 const verbose = yn(process.env.CANON_CMS_LOGGING);
@@ -33,115 +33,6 @@ const catcher = e => {
     console.error("Error in cmsRoute: ", e.message);
   }
   return [];
-};
-
-const profileReqFull = {
-  include: [
-    {association: "meta", separate: true},
-    {association: "content", separate: true},
-    {association: "generators", separate: true},
-    {association: "materializers", separate: true},
-    {association: "selectors", separate: true},
-    {
-      association: "sections", separate: true,
-      include: [
-        {association: "content", separate: true},
-        {association: "subtitles", include: [{association: "content", separate: true}], separate: true},
-        {association: "descriptions", include: [{association: "content", separate: true}], separate: true},
-        {association: "stats", include: [{association: "content", separate: true}], separate: true},
-        {association: "visualizations", separate: true},
-        {association: "selectors"}
-      ]
-    }
-  ]
-};
-
-const storyReqFull = {
-  include: [
-    {association: "content", separate: true},
-    {association: "authors", include: [{association: "content", separate: true}], separate: true},
-    {association: "descriptions", include: [{association: "content", separate: true}], separate: true},
-    {association: "footnotes", include: [{association: "content", separate: true}], separate: true},
-    {
-      association: "storysections", separate: true, 
-      include: [
-        {association: "content", separate: true},
-        {association: "subtitles", include: [{association: "content", separate: true}], separate: true},
-        {association: "descriptions", include: [{association: "content", separate: true}], separate: true},
-        {association: "stats", include: [{association: "content", separate: true}], separate: true},
-        {association: "visualizations", separate: true}
-      ]
-    }
-  ]
-};
-
-const sectionReqFull = {
-  include: [
-    {association: "content", separate: true},
-    {association: "subtitles", include: [{association: "content", separate: true}], separate: true},
-    {association: "descriptions", include: [{association: "content", separate: true}], separate: true},
-    {association: "stats", include: [{association: "content", separate: true}], separate: true},
-    {association: "visualizations", separate: true},
-    {association: "selectors"}
-  ]
-};
-
-const storysectionReqFull = {
-  include: [
-    {association: "content", separate: true},
-    {association: "subtitles", include: [{association: "content", separate: true}], separate: true},
-    {association: "descriptions", include: [{association: "content", separate: true}], separate: true},
-    {association: "stats", include: [{association: "content", separate: true}], separate: true},
-    {association: "visualizations", separate: true}
-  ]
-};
-
-/**
- * API paths are dynamically generated by folding over this list in the get/post methods that follow.
- * IMPORTANT: When new tables are added to the CMS, adding their exact tablename to this list will
- * automatically generate Create, Update, and Delete Routes (as specified later in the get/post methods)
- */
-const cmsTables = [
-  "author", "formatter", "generator", "materializer", "profile", "profile_meta",
-  "selector", "story", "story_description", "story_footnote", "storysection",
-  "storysection_description", "storysection_stat", "storysection_subtitle", "storysection_visualization",
-  "section", "section_description", "section_stat", "section_subtitle", "section_visualization", "section_selector"
-];
-
-/**
- * Some tables are translated to different languages using a corresponding "content" table, like "profile_content".
- * As such, some of the following functions need to take compound actions, e.g., insert a metadata record into
- * profile, THEN insert the "real" data into "profile_content." This list (subset of cmsTables) represents those
- * tables that need corresponding _content updates.
- */
-
-const contentTables = [
-  "author", "profile", "story", "story_description", "story_footnote", "storysection", "storysection_description",
-  "storysection_stat", "storysection_subtitle", "section", "section_description", "section_stat", "section_subtitle"
-];
-
-/**
- * Some tables need to know their own parents, for help with ordering. This lookup table allows
- * a given id to find its "siblings" and know where it belongs, ordering-wise
- */
-
-const parentOrderingTables = {
-  author: "story_id",
-  materializer: "profile_id",
-  profile_meta: "profile_id",
-  section: "profile_id",
-  section_description: "section_id",
-  section_selector: "section_id",
-  section_stat: "section_id",
-  section_subtitle: "section_id",
-  section_visualization: "section_id",
-  story_description: "story_id",
-  story_footnote: "story_id",
-  storysection: "story_id",
-  storysection_description: "storysection_id",
-  storysection_stat: "storysection_id",
-  storysection_subtitle: "storysection_id",
-  storysection_visualization: "storysection_id"
 };
 
 const sorter = (a, b) => a.ordering - b.ordering;
@@ -287,39 +178,6 @@ const duplicateSection = async(db, oldSection, pid, selectorLookup) => {
     }
   }
   return newSection.id;
-};
-
-/**
- * - contentArray of a cms entity
- * - db ref to process the upsert 
- * - config object with source, target, and variables
- * - req, to pass through to translateContent, so the API can path with origin/host
- * It updates the translation in-place, so has no return value.
- */
-const upsertTranslation = async(contentArray, db, ref, config, req) => {
-  const {source, target, variables} = config;
-  const defCon = contentArray.find(c => c.locale === source);
-  if (defCon) {
-    const formatterFunctions = await formatters4eval(db, source);
-    const {id, locale, ...content} = defCon; //eslint-disable-line
-    const translated = await translateContent(content, source, target, {variables, formatterFunctions}, req);
-    const newContent = {id, locale: target, ...translated};
-    await db[ref].upsert(newContent, {where: {id, locale: target}}).catch(catcher);
-  }
-};
-
-const translateSection = async(sid, config, db, req) => {
-  // Fetch the full section by the provided id, including content
-  const reqObj = Object.assign({}, sectionReqFull, {where: {id: sid}});
-  let section = await db.section.findOne(reqObj);
-  section = section.toJSON();
-  // Crawl through the object and translate each of its individual content rows, then update it in place
-  await upsertTranslation(section.content, db, "section_content", config, req);
-  for (const entityRef of ["subtitle", "description", "stat"]) {
-    for (const entity of section[`${entityRef}s`]) {
-      await upsertTranslation(entity.content, db, `section_${entityRef}_content`, config, req);
-    }
-  }
 };
 
 const pruneSearch = async(cubeName, dimension, levels, db) => {
