+++ packages/cms/src/api/translateRoute.js
@@ -18,7 +18,7 @@ module.exports = function(app) {
 
   app.post("/api/translatetest", async(req, res) => {
     const {text, target} = req.body; 
-    return res.json(`Would translate 7 ${target} ----> ${text}`);
+    return res.json(`Would translate 9 ${target} ----> ${text}`);
   });
   
 };
